requirements.txt
cython
gunicorn>=20.0.4
hemlock-demographics==0.0.5
hemlock-survey>=0.0.40
hemlock_demographics==0.0.4
psycopg2>=2.8.4
scikit-learn==0.24.0
scipy==1.5.4
setuptools